Adding products - create repository class - create productRepository class extending repository - change usersRepository class to extend repository
