+++ repositories/products.js
@@ -1,0 +1,7 @@
+const Repository = require('./repository');
+
+class ProductsRepository extends Repository{
+
+}
+
+module.exports = new ProductsRepository('products.json');
